Changes to calculate and log mismatching and missing tax codes
Itin vs response

+++ pe-engines-taxes/src/main/java/com/threevictors/aws/priceeye/exp/model/Taxes.java
@@ -2,13 +2,14 @@ package com.threevictors.aws.priceeye.exp.model;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 
 @Data
 public class Taxes {
     public String taxGroup;
     public String description;
-    public double taxAmount;
+    public BigDecimal taxAmount;
     public String taxAmountCurrency;
     public ArrayList<ChargeDetail> chargeDetails;
 }
